Add fixtures, commands, and utility functions for order management testing
- Created JSON fixtures for order cancellation, saving, and various order types (dine-in, delivery, takeaway).
- Implemented custom Cypress commands for user login, navigation, and page load waiting.
- Developed utility functions for order testing, including order existence checks and modal testing.
- Added migration to include average monthly rate in expense types.
- Updated Blade views for day toggle and global actions.
- Enhanced shift logs report page with detailed activity logging and context display.
- Implemented unit tests for product cost management service, covering various scenarios including cost updates and summaries.

+++ larament/database/migrations/2025_07_07_000004_create_inventory_items_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('inventory_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity')->default(0);
+            $table->decimal('quantity', 10, 2)->default(0);
             $table->timestamps();
         });
     }
